tests: reusable assertion for background filtering
